Refactored script to use gswin64c.exe location found by external function (#11)
The function was properly finding gswin64c.exe in a version folder properly. But the GhostScript calls of the rest of the script relied on `gswin64c` being added to the PATH variable or "shimmed" with shimgen.exe from Chocolatey.

There were certain areas of the script, such as `refreshenv` command that would cause the script to crash after, that might still need to be investigated.

Overall, this improvement works by not relying on any of that, but it might be worth it to still look into auto-adding to PATH or shimming in the future, to help out with troubleshooting ghostscript if you have to run the commands from the command-line yourself.