Bugfix: SETX PATH requires Admin (#11)
Refactored the way external function checks if gswin64c is added to PATH. Instead of failing out or relying completely on it, the script will now attempt to add it to PATH if it has been Run As Administrator, otherwise will continue with having found its full path and run the script. If the "debugging comments" are enabled, will display a Tip to Run the script as Administrator and will update the PATH for you.